get-data/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.3
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
boto3==1.26.25
    # via -r requirements.in
botocore==1.29.25
    # via
    #   boto3
    #   s3transfer
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
datasets==2.7.1
    # via -r requirements.in
dill==0.3.6
    # via
    #   datasets
    #   multiprocess
filelock==3.8.2
    # via huggingface-hub
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.11.0
    # via datasets
huggingface-hub==0.11.1
    # via datasets
idna==3.4
    # via
    #   requests
    #   yarl
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
loguru==0.6.0
    # via -r requirements.in
multidict==6.0.3
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.14
    # via datasets
numpy==1.23.5
    # via
    #   datasets
    #   pandas
    #   pyarrow
packaging==21.3
    # via
    #   datasets
    #   huggingface-hub
pandas==1.5.2
    # via datasets
pillow==9.3.0
    # via -r requirements.in
pyarrow==10.0.1
    # via datasets
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   botocore
    #   pandas
pytz==2022.6
    # via pandas
pyyaml==6.0
    # via
    #   datasets
    #   huggingface-hub
requests==2.28.1
    # via
    #   -r requirements.in
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   responses
responses==0.18.0
    # via datasets
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
tqdm==4.64.1
    # via
    #   datasets
    #   huggingface-hub
typing-extensions==4.4.0
    # via huggingface-hub
urllib3==1.26.13
    # via
    #   botocore
    #   requests
    #   responses
xxhash==3.1.0
    # via datasets
yarl==1.8.2
    # via aiohttp
